Player interactions.
Fixed the player being able to drop empty items on the ground.
Gave the player the ability to interact with objects. The interaction time is defined by the object being interacted with.
Reworked some of the player's logic to better fit with interactions.
Re-added "player_pick_up" control because Paul removed it ! bleh !